added some notes on what to work on
have to fix my boxplot, remove the outliers from it. The values above the max and below the min are outliers.

trhe comparison of the two tt races: can use find_peaks for the most obvious peaks instead of trying to find when oine race is faster than the other for all the peaks.