finished q2 and q3
I removed most of my progress for q2 because I intrepeted how the outliers were defined for the boxplot. I ended up using a predetermined value to find the oultiers.

for  q3, I compared the raw dfs and the smothed df.  We can see a drastic improvement in the clarity of the plot now. I've also used findpeaks to comare at different observcations which race will be faster.